refactor(app): move banner code to subfolder
