+++ lib/app/widgets/account_dialog.dart
@@ -5,8 +5,8 @@ import 'package:hive_cache/hive_cache.dart';
 import 'package:schulcloud/brand/brand.dart';
 import 'package:schulcloud/sign_in/sign_in.dart';
 
+import '../banner/service.dart';
 import '../data.dart';
-import '../services/banner.dart';
 import '../services/storage.dart';
 import '../utils.dart';
 import 'account_avatar.dart';
